recovery: Implement sysbar
Add a system bar (navigation bar) similar to the main Android system
with back and home buttons.  This makes it easier for users to figure
out how to go back on devices that lack hardware buttons, and also
provides a quick way to get back to the main menu.

Note only buttons that do not have a hardware equivalent are shown, in
order to prevent redundancy and confusion.

Change-Id: I7538749978837571a8c250c3c8e54ac127b39d84
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>